Redistribute project and add logger functionalities
The Hierarchy was reorganized following a model-view schema, more
suitable for a Qt project, instead model-view-controller. Thus, the
functionality of the controller was moved to the view package.

A logger was added in the main controller. It prints the log messages
to a QTextBrowser at the bottom of the interface, and the messages are
rendered depending on the logging level (debu/info/warning/error). Four
test buttons were added for checking the logger.

The interface design is now dynamic, and can be adapted to the total
size of the parent window.