Add overwrite device configuration functionality
The new functionality is implemented on the Device Manager pop-up
window. An open button allows to load an existing configuration, through
a file browser; and the save button override the new configuration.
Pop-up windows are also added, in order to prevent saving configurations
by mistake. In the future, a "delete configuration" button should be
added.

On the main window, the distribution of the devices buttons and boxes
has been updated. Up to 2 devices can be loaded, and a "remove button"
is created for each device. By default, their state is 'not connected'.