Add matplotlib graph to the interface
The graph is embedded in a vertical layout Qt object, and is formed by
a matplotlib canvas, inside of which there is a matplotlib figure and a
toolbar. A debug button is temporarily enabled for plotting a sample
line.

The connect buttons where improved. They try to connect to the device
and, if they can't, the socket instance is destroyed (actually the
assigned variable name is overridden) and a new one is created. This
should be improved in the future.
The load device button is also improved, checking if the selected device
actually exists (solving the situation where the configuration file is
deleted before pushing the button).