Set up routine for connecting with FPGA
The connection process includes an acknowledge instruction. The GUI
status labels and logger implement information about this procedure.

A dictionary for storing the available instrument commands is created,
containing by now only the ACK message.

a send_command function is added to the FreqMeterDevice class, in order
to add more abstraction to the functionality.

The 'connect' button allows now to disconnect the device as well, and
this can be repeated indefinitely. The 'remove device' button calls the
'disconnect' method before removing the device object.

Two 'fetch time' and two 'sample time' spin boxes have been added to the
GUI, a pair for each device.