Migrate project from PyQt4 to PyQt5
The changes are minor, as the QtDesigner tools allow to generate PyQt5
code from the designed interfaces (using pyuic5 instead pyuic4). Besides
generating again the base interfaces code, QWidgets submodule must now
be imported, as the buttons, scroll areas, message boxes... are children
of this module (Before they were children of QtGui).

Moreover, now it is used the Qt5 backednds from the matplotlib library
(The import lines were updated) instead the Qt4 ones.

Finally, the file path obtained with the open button in the device
manager is now stored at the first item of a list. Now it is also
checked if the cancel button was pressed.