Add routine to debug button for plotting up to 16 plots
The new functionality places as well a legend at the top of the graph.
Only the checked signals will be plotted, and this is updated every time
the fetch counter ends.

This is not definitive code, as it has to be properly modificated in
order to implement this functionality to the received real signals. By
now, only test values are plotted.