Remove redundant code
The remove and connect buttons were calling different methods, that only
differ on the GUI elements to interact with. These GUI elements were
grouped in 2-elements list, belonging the first element to the first
device, and the second element to the second device. Moreover, some
labels and buttons were renamed and a common style is used now.

The update_plots method has been adapted to the new data structure, but
it is only read the value of the first signal. Moreover, it is only
asking the external FPGA for the coarse measurement, independent of the
configured and checked signals.